+++ app/cloud/client/index.js
@@ -5,7 +5,7 @@ import './admin/cloudRegisterManually';
 import { BlazeLayout } from 'meteor/kadira:blaze-layout';
 
 import { registerAdminRoute } from '../../ui-admin/client/routes';
-import { AdminBox } from '../../ui-utils';
+import { AdminBox } from '../../ui-admin/client/AdminBox';
 import { hasAtLeastOnePermission } from '../../authorization';
 
 registerAdminRoute('/cloud', {
