+++ app/logger/client/viewLogs.js
@@ -2,7 +2,7 @@ import { BlazeLayout } from 'meteor/kadira:blaze-layout';
 import { Meteor } from 'meteor/meteor';
 import { Mongo } from 'meteor/mongo';
 
-import { AdminBox } from '../../ui-utils';
+import { AdminBox } from '../../ui-admin/client/AdminBox';
 import { hasAllPermission } from '../../authorization';
 import { t } from '../../utils';
 import { registerAdminRoute } from '../../ui-admin/client/routes';
