+++ app/ui-admin/client/adminFlex.js
@@ -5,7 +5,7 @@ import { Template } from 'meteor/templating';
 import { TAPi18n } from 'meteor/rocketchat:tap-i18n';
 
 import { settings } from '../../settings';
-import { menu, SideNav, AdminBox, Layout } from '../../ui-utils/client';
+import { menu, SideNav, Layout, AdminBox } from '../../ui-utils/client';
 import { t } from '../../utils';
 import { PrivateSettingsCachedCollection } from './SettingsCachedCollection';
 import { hasAtLeastOnePermission } from '../../authorization/client';
