+++ app/ui-utils/client/index.js
@@ -1,4 +1,4 @@
-export { AdminBox } from './lib/AdminBox';
+export { AdminBox } from '../../ui-admin/client/AdminBox';
 export { modal } from './lib/modal';
 export { SideNav } from './lib/SideNav';
 export { AccountBox } from './lib/AccountBox';
